Hoe irritant. Het verkrijgen van informatie vanuit een andere functie werkt nu, maar het parsen met PUG is nu weer een drama. En ik weiger om terug te gaan om te kijken hoe het zat ;)
Add download page and functionality

A new 'download' view has been added to the pug files. GetStudentFirstName function is now implemented in index.js to retrieve the student's first name based on the student number. If the student is not found, a 404 error is sent, otherwise the student's name is passed to the 'download' view.