FUCK YES! De studentnaam is nu terug te vinden in de waarde $value. #WINNEN!
Add nodemon and related dependencies to package-lock.json

In this update, the nodemon dependency was added to the package.json to enable auto-restart of the server whenever file changes are detected during development. In addition to this, associated dependencies required for nodemon to function efficiently were also included in the update.